# Static RTree for MultiPolygon Validation