## Description

Switches to using a static RTree for MultiPolygon validation.

I'm not expecting a performance change, this is more so that I can remove the Insert and Delete RTree methods.

## Check List

Have you:

- Added unit tests? N/A

- Add cmprefimpl tests? (if appropriate?) N/A

## Related Issue

- N/A

## Benchmark Results

There are some mild MultiPolygon perf improvements (unintended).

<details>

<summary>Click to expand</summary>

```
name                                                        old time/op    new time/op    delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
LineEnvelope/0-4                                              1.62ns ± 6%    1.62ns ± 6%     ~     (p=0.674 n=14+15)
LineEnvelope/1-4                                              1.07ns ± 6%    1.07ns ± 4%     ~     (p=0.757 n=15+14)
LineEnvelope/2-4                                              1.33ns ± 5%    1.33ns ± 3%     ~     (p=0.802 n=15+14)
LineEnvelope/3-4                                              1.19ns ± 5%    1.18ns ± 3%     ~     (p=0.938 n=15+14)
MarshalWKB/polygon/n=10-4                                      191ns ± 9%     187ns ± 7%     ~     (p=0.334 n=14+14)
MarshalWKB/polygon/n=100-4                                     679ns ±92%     590ns ±34%     ~     (p=0.650 n=13+12)
MarshalWKB/polygon/n=1000-4                                   3.96µs ±45%    5.43µs ±81%  +36.99%  (p=0.023 n=12+14)
MarshalWKB/polygon/n=10000-4                                  34.3µs ±42%    47.2µs ±64%     ~     (p=0.155 n=13+14)
UnmarshalWKB/polygon/n=10-4                                    319ns ± 7%     321ns ± 6%     ~     (p=0.524 n=15+12)
UnmarshalWKB/polygon/n=100-4                                   739ns ±38%     729ns ±11%     ~     (p=0.552 n=14+12)
UnmarshalWKB/polygon/n=1000-4                                 4.51µs ±31%   5.94µs ±147%     ~     (p=0.406 n=12+13)
UnmarshalWKB/polygon/n=10000-4                               53.2µs ±144%    52.2µs ±72%     ~     (p=0.847 n=14+15)
IntersectsLineStringWithLineString/n=10-4                     1.51µs ±14%    1.58µs ±25%     ~     (p=0.458 n=13+14)
IntersectsLineStringWithLineString/n=100-4                    21.5µs ± 8%    20.7µs ± 4%     ~     (p=0.107 n=14+11)
IntersectsLineStringWithLineString/n=1000-4                    207µs ± 6%     210µs ±13%     ~     (p=0.701 n=14+14)
IntersectsLineStringWithLineString/n=10000-4                  3.09ms ±11%    2.96ms ± 6%     ~     (p=0.065 n=15+13)
IntersectsMultiPointWithMultiPoint/n=20-4                     1.33µs ± 4%    1.34µs ± 5%     ~     (p=0.381 n=15+13)
IntersectsMultiPointWithMultiPoint/n=200-4                    14.1µs ± 3%    14.0µs ± 3%     ~     (p=0.156 n=15+13)
IntersectsMultiPointWithMultiPoint/n=2000-4                    142µs ± 5%     141µs ± 6%     ~     (p=0.331 n=14+15)
IntersectsMultiPointWithMultiPoint/n=20000-4                  1.50ms ± 3%    1.51ms ± 6%     ~     (p=0.983 n=15+14)
PolygonSingleRingValidation/n=10-4                            2.51µs ± 6%    2.54µs ±22%     ~     (p=0.765 n=14+13)
PolygonSingleRingValidation/n=100-4                           32.0µs ± 8%    34.4µs ±17%   +7.33%  (p=0.005 n=14+13)
PolygonSingleRingValidation/n=1000-4                           390µs ± 2%     397µs ± 6%     ~     (p=0.057 n=13+13)
PolygonSingleRingValidation/n=10000-4                         4.91ms ± 1%    5.01ms ± 8%     ~     (p=0.815 n=9+15)
PolygonMultipleRingsValidation/n=4-4                          6.75µs ± 4%    6.84µs ± 8%     ~     (p=0.347 n=12+14)
PolygonMultipleRingsValidation/n=36-4                         57.0µs ± 8%    56.7µs ± 6%     ~     (p=0.801 n=13+13)
PolygonMultipleRingsValidation/n=400-4                         739µs ± 6%     742µs ± 6%     ~     (p=0.705 n=12+14)
PolygonMultipleRingsValidation/n=4096-4                       9.23ms ±19%    8.65ms ± 4%     ~     (p=0.254 n=15+13)
PolygonZigZagRingsValidation/n=10-4                           11.2µs ± 7%    11.1µs ± 5%     ~     (p=0.839 n=14+14)
PolygonZigZagRingsValidation/n=100-4                           131µs ± 8%     129µs ± 6%     ~     (p=0.389 n=15+15)
PolygonZigZagRingsValidation/n=1000-4                         1.47ms ± 4%    1.47ms ± 3%     ~     (p=0.780 n=15+14)
PolygonZigZagRingsValidation/n=10000-4                        19.7ms ±12%    19.5ms ± 5%     ~     (p=0.880 n=14+15)
PolygonAnnulusValidation/n=10-4                               3.77µs ±12%    3.72µs ± 4%     ~     (p=0.762 n=13+13)
PolygonAnnulusValidation/n=100-4                              33.6µs ±13%    33.0µs ±11%     ~     (p=0.467 n=15+13)
PolygonAnnulusValidation/n=1000-4                              550µs ± 8%     557µs ± 4%     ~     (p=0.227 n=14+14)
PolygonAnnulusValidation/n=10000-4                            6.82ms ±20%    6.50ms ±12%     ~     (p=0.070 n=14+15)
MultipolygonValidation/n=1-4                                  450ns ±144%     400ns ±10%     ~     (p=0.104 n=15+15)
MultipolygonValidation/n=4-4                                   726ns ± 7%     916ns ±12%  +26.27%  (p=0.000 n=14+15)
MultipolygonValidation/n=16-4                                 4.20µs ± 5%    3.74µs ±14%  -10.97%  (p=0.000 n=12+13)
MultipolygonValidation/n=64-4                                 22.6µs ± 5%    16.9µs ±14%  -25.13%  (p=0.000 n=15+15)
MultipolygonValidation/n=256-4                                 164µs ±12%     105µs ± 6%  -35.76%  (p=0.000 n=14+15)
MultipolygonValidation/n=1024-4                                764µs ± 5%     496µs ± 6%  -35.12%  (p=0.000 n=15+15)
MultiPolygonTwoCircles/n=10-4                                6.96µs ±187%    3.90µs ±13%     ~     (p=0.448 n=13+13)
MultiPolygonTwoCircles/n=100-4                                38.5µs ±11%    44.2µs ±55%     ~     (p=0.097 n=13+14)
MultiPolygonTwoCircles/n=1000-4                                408µs ±10%     405µs ± 9%     ~     (p=0.650 n=12+13)
MultiPolygonTwoCircles/n=10000-4                              5.96ms ±17%    5.72ms ± 7%     ~     (p=0.217 n=15+13)
MultiPolygonMultipleTouchingPoints/n=1-4                      4.68µs ± 4%    4.81µs ±20%     ~     (p=0.296 n=14+13)
MultiPolygonMultipleTouchingPoints/n=10-4                     37.3µs ± 7%    37.8µs ± 9%     ~     (p=0.813 n=14+15)
MultiPolygonMultipleTouchingPoints/n=100-4                     446µs ± 9%     445µs ± 6%     ~     (p=0.960 n=13+13)
MultiPolygonMultipleTouchingPoints/n=1000-4                   5.03ms ± 3%    5.33ms ±18%   +5.89%  (p=0.000 n=13+13)
WKTParsing/point-4                                            1.71µs ±13%    1.71µs ±17%     ~     (p=0.659 n=13+14)
DistancePolygonToPolygonOrdering/n=100_swap=false-4           50.5µs ± 3%    53.1µs ±23%     ~     (p=0.548 n=12+15)
DistancePolygonToPolygonOrdering/n=100_swap=true-4            50.4µs ± 7%    51.2µs ±18%     ~     (p=0.943 n=13+14)
DistancePolygonToPolygonOrdering/n=1000_swap=false-4           751µs ± 5%     766µs ± 5%     ~     (p=0.112 n=14+15)
DistancePolygonToPolygonOrdering/n=1000_swap=true-4            750µs ± 4%     760µs ± 7%     ~     (p=0.362 n=13+13)
IntersectionPolygonWithPolygonOrdering/n=100_swap=false-4     5.49µs ± 8%    5.46µs ± 8%     ~     (p=0.755 n=15+12)
IntersectionPolygonWithPolygonOrdering/n=100_swap=true-4      5.61µs ± 8%    5.40µs ± 7%   -3.79%  (p=0.037 n=15+13)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=false-4    61.8µs ± 9%    59.8µs ± 9%     ~     (p=0.098 n=12+13)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=true-4     62.3µs ±21%    60.1µs ± 8%     ~     (p=0.231 n=14+12)
MultiLineStringIsSimpleManyLineStrings/n=100-4                54.6µs ±16%    52.9µs ± 9%     ~     (p=0.519 n=14+13)
MultiLineStringIsSimpleManyLineStrings/n=1000-4                624µs ±30%     556µs ± 5%  -10.92%  (p=0.019 n=15+15)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
IntersectionWithoutValidation/n=10-4                          49.0µs ± 9%    49.2µs ± 5%     ~     (p=0.616 n=15+15)
IntersectionWithoutValidation/n=100-4                         94.4µs ± 5%    93.9µs ± 6%     ~     (p=0.747 n=15+14)
IntersectionWithoutValidation/n=1000-4                         410µs ±13%     383µs ±20%   -6.48%  (p=0.026 n=15+14)
IntersectionWithoutValidation/n=10000-4                       3.51ms ±13%    3.56ms ± 7%     ~     (p=0.793 n=14+13)
NoOp/n=10-4                                                   3.83µs ± 6%    3.84µs ± 4%     ~     (p=0.691 n=15+14)
NoOp/n=100-4                                                  12.8µs ±11%    12.6µs ± 7%     ~     (p=0.967 n=15+15)
NoOp/n=1000-4                                                 89.2µs ± 3%    90.2µs ± 8%     ~     (p=0.595 n=15+15)
NoOp/n=10000-4                                                 953µs ±17%     930µs ±21%     ~     (p=0.591 n=14+15)
pkg:github.com/peterstace/simplefeatures/internal/perf goos:linux goarch:amd64
LineStringIsSimpleCircle/n=10-4                               2.09µs ± 8%    2.09µs ± 8%     ~     (p=0.802 n=13+14)
LineStringIsSimpleCircle/n=100-4                              31.3µs ±13%    31.2µs ± 9%     ~     (p=0.852 n=13+12)
LineStringIsSimpleCircle/n=1000-4                              379µs ± 3%     373µs ± 3%   -1.60%  (p=0.011 n=14+13)
LineStringIsSimpleCircle/n=10000-4                            4.80ms ± 5%    4.87ms ± 9%     ~     (p=0.603 n=14+14)
LineStringIsSimpleZigZag/10-4                                 1.92µs ± 6%    1.89µs ± 6%     ~     (p=0.331 n=13+14)
LineStringIsSimpleZigZag/100-4                                29.8µs ± 4%    29.7µs ± 5%     ~     (p=0.584 n=12+13)
LineStringIsSimpleZigZag/1000-4                                358µs ± 3%     358µs ± 4%     ~     (p=0.943 n=12+15)
LineStringIsSimpleZigZag/10000-4                              4.91ms ± 4%    4.87ms ± 6%     ~     (p=0.294 n=13+15)
SetOperation/n=4/Go_Intersection-4                            41.1µs ±13%    41.9µs ±10%     ~     (p=0.169 n=13+14)
SetOperation/n=4/Go_Difference-4                              42.6µs ± 7%    41.3µs ± 5%   -2.88%  (p=0.023 n=15+15)
SetOperation/n=4/Go_SymmetricDifference-4                     54.0µs ± 3%    54.7µs ± 7%     ~     (p=0.488 n=13+14)
SetOperation/n=4/Go_Union-4                                   43.1µs ± 4%    43.4µs ± 5%     ~     (p=0.519 n=14+13)
SetOperation/n=4/GEOS_Intersection-4                          49.8µs ± 5%    49.7µs ± 9%     ~     (p=0.652 n=14+15)
SetOperation/n=4/GEOS_Difference-4                            52.2µs ± 5%    51.7µs ± 4%     ~     (p=0.675 n=15+15)
SetOperation/n=4/GEOS_SymmetricDifference-4                   75.0µs ±14%    71.7µs ± 3%     ~     (p=0.080 n=15+13)
SetOperation/n=4/GEOS_Union-4                                 51.9µs ± 6%    51.7µs ± 5%     ~     (p=0.683 n=15+15)
SetOperation/n=8/Go_Intersection-4                            51.1µs ± 9%    51.8µs ±12%     ~     (p=0.717 n=15+13)
SetOperation/n=8/Go_Difference-4                              50.6µs ± 4%    50.8µs ± 5%     ~     (p=0.879 n=15+15)
SetOperation/n=8/Go_SymmetricDifference-4                     67.7µs ± 5%    67.0µs ± 4%     ~     (p=0.239 n=14+13)
SetOperation/n=8/Go_Union-4                                   52.2µs ± 8%    52.3µs ± 7%     ~     (p=0.622 n=12+13)
SetOperation/n=8/GEOS_Intersection-4                          57.4µs ± 4%    57.5µs ± 5%     ~     (p=0.946 n=14+14)
SetOperation/n=8/GEOS_Difference-4                            58.4µs ± 5%    58.3µs ± 5%     ~     (p=0.806 n=15+15)
SetOperation/n=8/GEOS_SymmetricDifference-4                   84.0µs ± 4%    86.3µs ± 5%   +2.71%  (p=0.002 n=13+14)
SetOperation/n=8/GEOS_Union-4                                 59.3µs ± 8%    59.2µs ± 4%     ~     (p=0.867 n=14+13)
SetOperation/n=16/Go_Intersection-4                           70.3µs ±12%    69.8µs ± 8%     ~     (p=0.583 n=13+14)
SetOperation/n=16/Go_Difference-4                             73.6µs ± 6%    72.5µs ±10%     ~     (p=0.217 n=15+15)
SetOperation/n=16/Go_SymmetricDifference-4                     107µs ±11%     103µs ± 4%     ~     (p=0.098 n=15+15)
SetOperation/n=16/Go_Union-4                                  76.8µs ± 3%    76.5µs ± 4%     ~     (p=0.683 n=15+15)
SetOperation/n=16/GEOS_Intersection-4                         63.1µs ± 4%    63.1µs ± 4%     ~     (p=0.905 n=13+14)
SetOperation/n=16/GEOS_Difference-4                           69.1µs ± 4%    69.3µs ± 5%     ~     (p=0.683 n=14+15)
SetOperation/n=16/GEOS_SymmetricDifference-4                   109µs ± 7%     109µs ± 6%     ~     (p=0.902 n=15+15)
SetOperation/n=16/GEOS_Union-4                                71.5µs ± 6%    73.8µs ±10%     ~     (p=0.201 n=14+15)
SetOperation/n=32/Go_Intersection-4                            119µs ± 7%     120µs ± 3%     ~     (p=0.128 n=14+13)
SetOperation/n=32/Go_Difference-4                              124µs ± 8%     124µs ± 3%     ~     (p=0.905 n=14+13)
SetOperation/n=32/Go_SymmetricDifference-4                     174µs ± 8%     175µs ± 7%     ~     (p=0.650 n=15+13)
SetOperation/n=32/Go_Union-4                                   129µs ± 6%     131µs ± 8%     ~     (p=0.511 n=14+14)
SetOperation/n=32/GEOS_Intersection-4                         75.9µs ± 3%    76.1µs ± 5%     ~     (p=0.724 n=13+13)
SetOperation/n=32/GEOS_Difference-4                           83.8µs ± 4%    84.8µs ± 6%     ~     (p=0.285 n=14+14)
SetOperation/n=32/GEOS_SymmetricDifference-4                   145µs ±11%     148µs ± 9%     ~     (p=0.234 n=15+14)
SetOperation/n=32/GEOS_Union-4                                90.2µs ± 4%    92.6µs ±11%     ~     (p=0.217 n=14+15)
SetOperation/n=64/Go_Intersection-4                            202µs ± 8%     200µs ±10%     ~     (p=0.436 n=15+15)
SetOperation/n=64/Go_Difference-4                              216µs ±12%     219µs ± 7%     ~     (p=0.290 n=14+15)
SetOperation/n=64/Go_SymmetricDifference-4                     304µs ± 6%     305µs ± 6%     ~     (p=0.967 n=15+15)
SetOperation/n=64/Go_Union-4                                   229µs ± 8%     225µs ± 5%     ~     (p=0.259 n=13+14)
SetOperation/n=64/GEOS_Intersection-4                          100µs ± 5%      98µs ± 2%     ~     (p=0.103 n=15+12)
SetOperation/n=64/GEOS_Difference-4                            122µs ± 6%     121µs ± 4%     ~     (p=0.425 n=15+14)
SetOperation/n=64/GEOS_SymmetricDifference-4                   223µs ±10%     225µs ±10%     ~     (p=0.635 n=14+14)
SetOperation/n=64/GEOS_Union-4                                 132µs ± 3%     132µs ± 3%     ~     (p=0.914 n=15+14)
SetOperation/n=128/Go_Intersection-4                           367µs ± 4%     382µs ±16%     ~     (p=0.160 n=12+14)
SetOperation/n=128/Go_Difference-4                             386µs ± 6%     397µs ±10%   +2.92%  (p=0.029 n=13+14)
SetOperation/n=128/Go_SymmetricDifference-4                    555µs ±16%     539µs ± 7%     ~     (p=0.804 n=14+14)
SetOperation/n=128/Go_Union-4                                  410µs ± 9%     400µs ± 3%     ~     (p=0.081 n=13+13)
SetOperation/n=128/GEOS_Intersection-4                         150µs ± 3%     151µs ± 5%     ~     (p=0.511 n=14+14)
SetOperation/n=128/GEOS_Difference-4                           179µs ± 8%     180µs ± 7%     ~     (p=0.488 n=14+13)
SetOperation/n=128/GEOS_SymmetricDifference-4                  346µs ± 4%     341µs ± 4%     ~     (p=0.053 n=15+12)
SetOperation/n=128/GEOS_Union-4                                198µs ± 4%     198µs ± 6%     ~     (p=0.806 n=15+15)
SetOperation/n=256/Go_Intersection-4                           662µs ± 7%     671µs ±10%     ~     (p=0.402 n=13+14)
SetOperation/n=256/Go_Difference-4                             748µs ±10%     727µs ± 9%     ~     (p=0.280 n=13+14)
SetOperation/n=256/Go_SymmetricDifference-4                   1.11ms ±21%    1.06ms ± 8%     ~     (p=0.440 n=15+13)
SetOperation/n=256/Go_Union-4                                  803µs ±21%     754µs ± 4%     ~     (p=0.185 n=14+13)
SetOperation/n=256/GEOS_Intersection-4                         226µs ± 6%     222µs ± 3%     ~     (p=0.056 n=14+14)
SetOperation/n=256/GEOS_Difference-4                           297µs ± 5%     291µs ± 1%   -2.06%  (p=0.009 n=14+11)
SetOperation/n=256/GEOS_SymmetricDifference-4                  668µs ±10%     649µs ± 9%     ~     (p=0.128 n=13+14)
SetOperation/n=256/GEOS_Union-4                                334µs ± 5%     330µs ± 7%     ~     (p=0.302 n=13+14)
SetOperation/n=512/Go_Intersection-4                          1.36ms ±10%    1.37ms ±10%     ~     (p=0.571 n=14+14)
SetOperation/n=512/Go_Difference-4                            1.43ms ± 4%    1.46ms ± 9%     ~     (p=0.142 n=13+15)
SetOperation/n=512/Go_SymmetricDifference-4                   2.02ms ± 9%    2.07ms ± 7%     ~     (p=0.118 n=15+13)
SetOperation/n=512/Go_Union-4                                 1.51ms ± 9%    1.52ms ±11%     ~     (p=0.650 n=14+13)
SetOperation/n=512/GEOS_Intersection-4                         403µs ± 5%     397µs ± 5%     ~     (p=0.134 n=14+15)
SetOperation/n=512/GEOS_Difference-4                           527µs ± 4%     532µs ± 5%     ~     (p=0.352 n=14+14)
SetOperation/n=512/GEOS_SymmetricDifference-4                 1.18ms ± 5%    1.17ms ± 3%     ~     (p=0.793 n=14+13)
SetOperation/n=512/GEOS_Union-4                                596µs ± 8%     600µs ± 6%     ~     (p=0.561 n=14+15)
SetOperation/n=1024/Go_Intersection-4                         2.73ms ±25%    2.59ms ± 8%     ~     (p=0.234 n=15+14)
SetOperation/n=1024/Go_Difference-4                           2.84ms ± 6%    2.81ms ± 5%     ~     (p=0.311 n=13+13)
SetOperation/n=1024/Go_SymmetricDifference-4                  4.17ms ± 5%    4.13ms ± 6%     ~     (p=0.603 n=14+14)
SetOperation/n=1024/Go_Union-4                                2.98ms ± 7%    2.95ms ± 5%     ~     (p=0.477 n=15+14)
SetOperation/n=1024/GEOS_Intersection-4                        739µs ± 4%     741µs ± 4%     ~     (p=0.870 n=15+15)
SetOperation/n=1024/GEOS_Difference-4                         1.05ms ± 6%    1.07ms ± 7%     ~     (p=0.164 n=14+14)
SetOperation/n=1024/GEOS_SymmetricDifference-4                2.36ms ± 4%    2.40ms ± 8%     ~     (p=0.205 n=12+13)
SetOperation/n=1024/GEOS_Union-4                              1.23ms ± 5%    1.23ms ± 4%     ~     (p=0.860 n=14+12)
SetOperation/n=2048/Go_Intersection-4                         5.64ms ± 5%    5.60ms ±13%     ~     (p=0.320 n=12+13)
SetOperation/n=2048/Go_Difference-4                           6.19ms ±19%    5.96ms ±11%     ~     (p=0.239 n=14+13)
SetOperation/n=2048/Go_SymmetricDifference-4                  8.49ms ±12%    8.28ms ± 6%     ~     (p=0.274 n=15+13)
SetOperation/n=2048/Go_Union-4                                6.24ms ± 5%    6.63ms ±30%     ~     (p=0.265 n=14+14)
SetOperation/n=2048/GEOS_Intersection-4                       1.48ms ± 6%    1.47ms ± 2%     ~     (p=0.555 n=15+13)
SetOperation/n=2048/GEOS_Difference-4                         1.89ms ± 3%    1.88ms ± 3%     ~     (p=0.503 n=13+12)
SetOperation/n=2048/GEOS_SymmetricDifference-4                4.63ms ±11%    4.62ms ± 7%     ~     (p=0.345 n=15+15)
SetOperation/n=2048/GEOS_Union-4                              2.27ms ± 9%    2.26ms ± 4%     ~     (p=0.533 n=14+15)
SetOperation/n=4096/Go_Intersection-4                         11.6ms ±12%    11.6ms ± 6%     ~     (p=0.943 n=13+14)
SetOperation/n=4096/Go_Difference-4                           13.2ms ±13%    12.2ms ± 6%   -7.39%  (p=0.000 n=14+14)
SetOperation/n=4096/Go_SymmetricDifference-4                  18.2ms ±15%    17.6ms ± 7%     ~     (p=0.141 n=14+13)
SetOperation/n=4096/Go_Union-4                                13.8ms ± 9%    13.6ms ±12%     ~     (p=0.512 n=15+15)
SetOperation/n=4096/GEOS_Intersection-4                       2.75ms ± 6%    2.70ms ± 5%     ~     (p=0.102 n=15+14)
…